userService/be/requirements.txt
django==4.1.6
psycopg2==2.9.5
model-bakery==1.10.1
djangorestframework==3.14.0
black==23.11.0
django-cors-headers==3.14.0
pika==1.3.2
